#14 Checking length for servicename in deleteServiceRequest

+++ football_marketplace-app/src/js/dbqueries.js
@@ -160,11 +160,14 @@ const getUsername = async (address)=>{
 const deleteServiceRequest = (address,service_name)=>{
     //TODO: Check if address does not contain malicious code https://github.com/swisstackle/football_marketplace/issues/34
     service_name = encodeURIComponent(service_name);
-    client.query('DELETE FROM service_requests WHERE address=$1 AND service_name=$2',[address,service_name],(error) =>{
-        if(error){
-            throw error;
-        }
-    })
+    if(service_name.length > 0){
+        client.query('DELETE FROM service_requests WHERE address=$1 AND service_name=$2',[address,service_name],(error) =>{
+            if(error){
+                throw error;
+            }
+        })
+    }
+
 }
 const deleteService = (address, name)=>{
     //TODO: Check if address does not contain malicious code https://github.com/swisstackle/football_marketplace/issues/34
